runtime: Rename the rpc crate to enclave_rpc
