+++ .changelog/2469.internal.1.md
@@ -1,0 +1,1 @@
+runtime: Rename the rpc crate to enclave_rpc
