+++ keymanager-lib/src/kdf.rs
@@ -27,8 +27,8 @@ use oasis_core_runtime::{
         runtime::RuntimeId,
         sgx::egetkey::egetkey,
     },
+    enclave_rpc::Context as RpcContext,
     executor::Executor,
-    rpc::Context as RpcContext,
     runtime_context,
     storage::StorageContext,
     BUILD_INFO,
